Created README.md
Overview of all present classes exept Cubemap class because it's not working yet but the idea is to make a colorful and dynamic skybox to create aurora effects around the environment and heaven full of shining stars in the future. 

It could be utilized as cubes with dynamic colors for building stuff too.